requirements.txt
gradio>=4.0.0,<5.0.0
torch>=2.1.0
torchvision>=0.16.0
pillow>=10.0.0
